env: sunxi: Update default env fat device
Usually eMMC is default env fat device for environment,
if MMC_SUNXI_SLOT_EXTRA != 1 Sunxi always probed emmc
device as 1. but with DM_MMC it can be more possible to
probe eMMC as device 2 since for most of the sunxi platforms
eMMC is configured mmc2.

So update the env default device as 2 if DM_MMC and
MMC_SUNXI_SLOT_EXTRA != 1 but some boards may not use all
possible mmc devices or partly disabled in DT, for those update
the device in board specific defconfig.

Cc: Olliver Schinagl <oliver@schinagl.nl>
Cc: Hans de Goede <hdegoede@redhat.com>
Cc: Marcus Cooper <codekipper@gmail.com>
Cc: Stefan Mavrodiev <stefan@olimex.com>
Cc: Paul Kocialkowski <contact@paulk.fr>
Cc: Chen-Yu Tsai <wens@csie.org>
Signed-off-by: Jagan Teki <jagan@amarulasolutions.com>