roc-rk3399-pc: Set low power leds, power key only if POR
ROC-RK3399-PC has specific set of configurations for
on-board led setup.

Due to easiness for user to know the state of the board
roc-rk339-pc board code will setup the low power led
on/off, and waiting for user to press power key and then
glow full power led.

All this needs to happen only during power-on-reset not
for soft reset or WDT. So add reset cause check to POR
and configure the low power leds, power key only. Note
that the glowing red led is common across any reset.

Reported-by: Markus Reichl <m.reichl@fivetechno.de>
Signed-off-by: Jagan Teki <jagan@amarulasolutions.com>
Tested-by: Suniel Mahesh <sunil@amarulasolutions.com>