roc-rk3399-pc: Support pwr_key env for low power
Set low power leds, power key only if POR, but it has
an issue for remote users to press the power key to
continue boot.

So, add env variable pwr_key via SPI flash so-that if
pwr_key is 'y' then user interaction of pressing the
power key would require otherwise not.

Signed-off-by: Jagan Teki <jagan@amarulasolutions.com>